Performed first query
Frontend and backend are now talking to each other. When the DefaultDisplay component is rendered, it only makes one query so far that grabs the first five seconds of the data. More work needs to be done to implement pagination. Also changed sample_chart.js to DefaultDisplay.js. Created the component AllPatients.js to display in the second menu item centercomponent and moved createUser to the third centercomponent if clause in dashboard